requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
comtypes==1.1.14
emoji==2.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isOdd==0.1.2
langdetect==1.0.9
progress==1.6
pydub==0.25.1
pypiwin32==223
pyTelegramBotAPI==4.10.0
python-telegram-bot==20.0
pyttsx3==2.90
pywin32==305
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
telebot==0.0.4
urllib3==1.26.14
